Adjust the flicker values.
We're using inverted mode on all of the outputs (0 = on, 255 = off)
because then we have consistent inputs on all of the ports, including
the timer interrupt.

This means that 135+random() actually makes the lights half-dim and then
dims the more, which isn't what we want.

Instead just use a random value from full brightness to almost dark.